add an attempt for codegen_mstore8
the values are stored in memory but not in the right addresses. It looks like for some reason the offset is being multiplied by 8, because when the offset is 1 the value ends up in the 8th byte, and when the offset is 2, it ends up un the 16th byte and so on. The same thing happens with MSTORE, but MSTORE has another issue in which the most significative zeroes are not stored