Implement default for parsed account info from ef tests
This is necessary because some tests don't provide the required "to" address pointed by the transaction, as they are expected to be invalid and don't actually require the account. However, the parsing of the "to" addresses is done before the validation of the transaction, resulting in a panic in `ef_tests.rs`, line 193. To fix this, a default with valid `code`, (starting with `0x`) is required. This allows tests that expect to fail to be correctly parsed before being executed.